requirements.txt
Pillow
requests
openai
